requirements.txt
requests==2.27.1
python-dotenv==0.20.0
python-telegram-bot==13.11
redis==3.2.1
yandex-geocoder==2.0.0
geopy==2.2.0
Flask==1.0
gunicorn==19.6.0